Add response to deleteCartItem

+++ src/main/java/spring/practice/elmenus_lite/controller/CartItemController.java
@@ -4,8 +4,10 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import spring.practice.elmenus_lite.dto.BaseResponse;
+import spring.practice.elmenus_lite.dto.CartDto;
 import spring.practice.elmenus_lite.dto.CartItemDto;
 import spring.practice.elmenus_lite.dto.UpdateItemQuantityRequestBody;
+import spring.practice.elmenus_lite.handlerException.NotFoundCustomException;
 import spring.practice.elmenus_lite.service.CartItemService;
 
 @RestController
@@ -36,10 +38,20 @@ public class CartItemController {
         return ResponseEntity.ok(updatedItem);
     }
 
+
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteCartItem(@PathVariable Long id) {
-        cartItemService.deleteCartItemById(id);
-        return ResponseEntity.noContent().build();
+    public ResponseEntity<BaseResponse<CartItemDto>> deleteCartItem(@PathVariable Long id) {
+        if (id <= 0) {
+            return ResponseEntity.badRequest().body(new BaseResponse<>(false, "Invalid Cart Item ID", null));
+        }
+        try {
+            cartItemService.deleteCartItemById(id);
+            return ResponseEntity.ok(new BaseResponse<>(true, "Cart item deleted successfully", null));
+        } catch (NotFoundCustomException ex) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new BaseResponse<>(false, ex.getMessage(), null));
+        } catch (Exception e){
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(new BaseResponse<>(false, e.getMessage(), null));
+        }
     }
 
     @PatchMapping("/{cartId}/{cartItemId}/quantity")
